previousCommand BackGUI.java
Added a new check, so if the player executed a command to get to the current GUI, they will be sent there. Otherwise, they'll be sent to their previousInventory.

This is to prevent errors in the case that IridiumSkyblock is being intertwined with another GUI plugin, to make more advanced GUIs.